Report controller merged with Entry controller
